Got the contact page to work with the new routes system.
