+++ routes/admin/downloads.js
@@ -1,6 +1,6 @@
 const routes = require('express').Router();
 
-routes.post('/', (request, result) =>
+routes.get('/', (request, result) =>
 {
 
 });
