+++ routes/downloads/index.js
@@ -1,8 +1,12 @@
 const routes = require('express').Router();
 
-routes.post('/', (request, result) =>
-{
+/** used to parse the request URL */
+const url = require('url');
 
+routes.get('/', (request, result) =>
+{
+    const filename = url.parse(request.url, true).pathname;
+    require("../../includes/downloads.js").main(result, filename);
 });
