+++ routes/index.js
@@ -6,12 +6,17 @@ routes.use('/about', about);
 const contact = require('./contact');
 routes.use('/contact', contact);
 
+const includes = require('./includes');
+routes.use('/includes', includes);
+
 const projects = ["/steam/"];
+
+
 
 
 routes.get('/', (request, result) =>
 {
-    //blog home
+    //blog home page
 });
